Create Compose Theme Related Classes
Specify Colors in the ApplicationTheme for use in the entire app. Furthermore, create a CustomColorsPalette to handle custom red (for debit) and green (for credit) colors in both Light and Dark Modes.

Finally, create a @CombinedThemePreviews annotation to help display Composable Previews in Light and Dark Mode in one declaration.